+++ example/lib/example5_themes_playground/widgets/pages/theme_topic.dart
@@ -69,7 +69,7 @@ const List<ThemeTopic> themeTopics = <ThemeTopic>[
   ),
   ThemeTopic(
     heading: 'AppBar and BottomAppBar',
-    buttonLabel: 'AppBar\nBottomAppBar',
+    buttonLabel: 'AppBar',
     icon: Icons.web_asset_outlined,
   ),
   ThemeTopic(